moved tests to separate folders
