+++ index.js
@@ -7,9 +7,9 @@ import handleValidationErrors from './utils/handleValidationErrors.js';
 import { loginValidation, registerValidation } from './validations/auth.js';
 import { groupCreateValidation } from './validations/group.js';
 import { taskValidation } from './validations/task.js';
-import * as UserController from './controllers/UserController.js';
-import * as GroupController from './controllers/GroupController.js';
-import * as TaskController from './controllers/TaskController.js'
+import * as UserController from './controllers/UserController/UserController.js';
+import * as GroupController from './controllers/GroupController/GroupController.js';
+import * as TaskController from './controllers/TaskController/TaskController.js'
 
 dotenv.config();
 const app = express();
